FarmManagementApp/backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic[email]==1.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.7.1
psycopg2-binary==2.9.1
python-multipart==0.0.5
python-dotenv==0.19.0
geoalchemy2==0.9.0
shapely==1.7.1
fastapi-cache2[redis]==0.1.8